LayoutTests/fast/css3-text/css3-text-decoration/text-decoration-skip/text-decoration-skip-roundtrip.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../../resources/js-test-pre.js"></script>
</head>
<body>
  <script type="text/javascript">
    function testInkIsValid(stylesheet) {
      cssRule = stylesheet.cssRules.item(0);
      shouldBe("cssRule.type", "cssRule.STYLE_RULE");
      declaration = cssRule.style;
      shouldBe("declaration.length", "1");
      shouldBe("declaration.getPropertyValue('-webkit-text-decoration-skip')", "\"ink\"");
      stylesheet.deleteRule(0);
    }

    function testInvalidRule(stylesheet) {
      cssRule = stylesheet.cssRules.item(0);
      shouldBe("cssRule.type", "cssRule.STYLE_RULE");
      declaration = cssRule.style;
      shouldBe("declaration.length", "0");
      stylesheet.deleteRule(0);
    }

    function testNoneIsValid(stylesheet) {
      cssRule = stylesheet.cssRules.item(0);
      shouldBe("cssRule.type", "cssRule.STYLE_RULE");
      declaration = cssRule.style;
      shouldBe("declaration.length", "1");
      shouldBe("declaration.getPropertyValue('-webkit-text-decoration-skip')", "\"none\"");
      stylesheet.deleteRule(0);
    }

    var styleElement = document.createElement("style");
    document.head.appendChild(styleElement);
    stylesheet = styleElement.sheet;

    stylesheet.insertRule(".p { -webkit-text-decoration-skip: ink; }", 0);
    testInkIsValid(stylesheet);

    stylesheet.insertRule(".p { -webkit-text-decoration-skip: ink ink ink ink ink; }", 0);
    testInkIsValid(stylesheet);

    stylesheet.insertRule(".p { -webkit-text-decoration-skip: garbage; }", 0);
    testInvalidRule(stylesheet);

    stylesheet.insertRule(".p { -webkit-text-decoration-skip: garbage ink; }", 0);
    testInkIsValid(stylesheet);

    stylesheet.insertRule(".p { -webkit-text-decoration-skip: ink garbage; }", 0);
    testInkIsValid(stylesheet);
    
    stylesheet.insertRule(".p { -webkit-text-decoration-skip: garbage ink garbage; }", 0);
    testInkIsValid(stylesheet);
    
    stylesheet.insertRule(".p { -webkit-text-decoration-skip: ink garbage ink; }", 0);
    testInkIsValid(stylesheet);
    
    stylesheet.insertRule(".p { -webkit-text-decoration-skip: none; }", 0);
    testNoneIsValid(stylesheet);
    
    stylesheet.insertRule(".p { -webkit-text-decoration-skip: ; }", 0);
    testInvalidRule(stylesheet);
  </script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
