LayoutTests/fast/css3-text/css3-text-decoration/text-decoration-skip/text-decoration-skip-roundtrip-expected.txt

PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-text-decoration-skip') is "ink"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-text-decoration-skip') is "ink"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-text-decoration-skip') is "ink"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-text-decoration-skip') is "ink"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-text-decoration-skip') is "ink"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-text-decoration-skip') is "ink"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-text-decoration-skip') is "none"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
